Refactoring: submodules + wrap common code in new functions
