+++ api/src/rust_jni/exception.rs
@@ -1,4 +1,4 @@
-use jni::{JNIEnv, sys::jobject};
+use super::*;
 use std::{error::Error, any::Any};
 
 pub const JNI_NULL: jobject = std::ptr::null::<jobject>() as jobject;
@@ -52,6 +52,28 @@ macro_rules! throw_and_exit {
         _throw_inner($env, $exception, $description);
         panic!("Thrown exception: {} for reason: {}", $exception, $description)
     };
+}
+
+#[macro_export]
+/// Map a type Result<T, E> to a jobject/jboolean if Ok(),
+/// otherwise throw exception and return default.
+macro_rules! map_or_throw {
+
+    // Return type: jobject
+    ($env:expr, $result: expr, $obj_path: expr, $exception:expr, $description:expr) => {{
+        ($result).map_or_else(
+            |e| throw!(&$env, $exception, format!("{:?}: {:?}", $description, e).as_str(), JNI_NULL),
+            |ret| *return_jobject(&$env, ret, $obj_path)
+        )
+    }};
+
+    // Return type: jboolean
+    ($env:expr, $result:expr, $exception:expr, $description:expr) => {{
+        ($result).map_or_else(
+            |e| throw!(&$env, $exception, format!("{:?}: {:?}", $description, e).as_str(), JNI_FALSE),
+            |ret| if ret { JNI_TRUE } else { JNI_FALSE }
+        )
+    }};
 }
 
 /// Transform a function into an implementation of a Java side native function.
