+++ jni/src/main/java/com/horizen/librustsidechains/Library.java
@@ -126,7 +126,7 @@ public class Library {
     // Used for testing purposes
     private static native void nativePanickingFunction();
 
-    public static void panickingFunction() {
+    protected static void panickingFunction() {
         nativePanickingFunction();
     }
